move list dict and named tuple tests out of py3 and into test_list_dict.py
